Move et Jump : actions de l'automate et valeurs par défaut implémentées

+++ project.game/src/Model/Entity.java
@@ -102,14 +102,14 @@ public abstract class Entity
     m_isMoving = true;
   }
   
-  public void doMove( Direction dir, int time )
+  public void doMove( Direction dir, long time )
   {
     m_moveDirection = dir.toAngle( m_orientation );
     m_timeToMove = time;
     m_isMoving = true;
   }
   
-  public void doMove( double dir, int time )
+  public void doMove( double dir, long time )
   {
     m_moveDirection = dir;
     m_timeToMove = time;
